requirements.txt
beautifulsoup4==4.12.2
requests==2.32.0
pandas==2.2.2
pyyaml==6.0
spacy==3.7.1
scikit-learn==1.3.0
nltk==3.8.1
tqdm==4.66.1
python-dotenv==1.0.0
pytrends==4.9.2